requirements.txt
aiofiles
jinja2
starlette
uvicorn
pyaml
databases
sqlalchemy
asyncpg
psycopg2

